// src/app/interfaces/IElements.ts
export interface IElements {
  Artikelname: string;
  Bein: string;
  Beschreibung: string;
  Bildname: string;
  Geschlecht: string;
  Grammatur: string;
  Hauptartikelnr: string;
  Hersteller: string;
  Herstellung: string;
  Kragen: string;
  Material: string;
  Materialangaben: string;
  Produktart: string;
  Taschenart: string;
  Ursprungsland: string;
  Id?: number;
  Aermel: string;
  action?: any;
}